avances Issue #26
